+++ src/main.ts
@@ -1,6 +1,6 @@
 import {createApp} from "vue";
 import App from "./App.vue";
-import router from "./router/router";
+import router from "./pages/router";
 import "intro.js/introjs.css"
 import {createPinia} from "pinia";
 import Vue3Storage from "vue3-storage";
