+++ src/pages/article-copy-tool/components/InputFuncButtons.vue
