+++ src/pages/article-copy-tool/components/OutputArea.vue
