+++ src/pages/article-copy-tool/components/OutputFuncButtons.vue
