+++ src/pages/article-copy-tool/components/PreferenceConfig.vue
