+++ src/pages/components/.gitattributes
