+++ src/pages/components/CopyrightFooter.vue
